+++ support-lib/jni/Marshal.hpp
@@ -19,7 +19,6 @@
 #pragma once
 
 #include "djinni_support.hpp"
-#include "../cpp/Future.hpp"
 #include <array>
 #include <cassert>
 #include <chrono>
@@ -813,122 +812,6 @@ namespace djinni
     public:
         static jdoubleArray makePrimitiveArray(JNIEnv* jniEnv, jsize size) {
             return jniEnv->NewDoubleArray(size);
-        }
-    };
-
-    struct PromiseJniInfo {
-        const GlobalRef<jclass> clazz { jniFindClass("com/snapchat/djinni/Promise") };
-        const jmethodID constructor { jniGetMethodID(clazz.get(), "<init>", "()V") };
-        const jmethodID method_get_future { jniGetMethodID(clazz.get(), "getFuture", "()Lcom/snapchat/djinni/Future;") };
-        const jmethodID method_set_value { jniGetMethodID(clazz.get(), "setValue", "(Ljava/lang/Object;)V") };
-        const jmethodID method_set_exception { jniGetMethodID(clazz.get(), "setException", "(Ljava/lang/Throwable;)V") };
-    };
-
-    struct FutureJniInfo {
-        const GlobalRef<jclass> clazz { jniFindClass("com/snapchat/djinni/Future") };
-        const jmethodID method_then { jniGetMethodID(clazz.get(), "then", "(Lcom/snapchat/djinni/Future$FutureHandler;)Lcom/snapchat/djinni/Future;") };
-    };
-
-    struct NativeFutureHandlerJniInfo {
-        const GlobalRef<jclass> clazz { jniFindClass("com/snapchat/djinni/NativeFutureHandler") };
-        const jmethodID constructor { jniGetMethodID(clazz.get(), "<init>", "(JJ)V") };
-    };
-
-    struct RuntimeExceptionJniInfo {
-        const GlobalRef<jclass> clazz { jniFindClass("java/lang/RuntimeException") };
-        const jmethodID constructor { jniGetMethodID(clazz.get(), "<init>", "(Ljava/lang/String;)V") };
-    };
-
-    struct ThrowableJniInfo {
-        const GlobalRef<jclass> clazz { jniFindClass("java/lang/Throwable") };
-        const jmethodID method_get_message { jniGetMethodID(clazz.get(), "getMessage", "()Ljava/lang/String;") };
-    };
-
-    using NativeFutureHandlerFunc = void (*)(JNIEnv* jniEnv, jlong nativePromise, jobject jres, jthrowable jex);
-
-    template <class RESULT>
-    class FutureAdaptor
-    {
-        using CppResType = typename RESULT::CppType;
-
-        struct PromiseHolder {
-            GlobalRef<jobject> promise;
-
-            void set(JNIEnv * env, jobject localRef) {
-                promise = GlobalRef<jobject>(env, localRef);
-            }
-            auto get() const { return promise.get(); }
-        };
-    public:
-        using CppType = Future<CppResType>;
-        using JniType = jobject;
-
-        using Boxed = FutureAdaptor;
-
-        static CppType toCpp(JNIEnv* jniEnv, JniType j)
-        {
-            using NativePromiseType = Promise<CppResType>;
-            
-            auto p = std::make_unique<NativePromiseType>();
-            auto f = p->getFuture();
-
-            NativeFutureHandlerFunc FutureHandler = [] (JNIEnv* jniEnv, jlong nativePromise, jobject jres, jthrowable jex) {
-                std::unique_ptr<NativePromiseType> promise {
-                    reinterpret_cast<NativePromiseType*>(nativePromise)
-                };
-                if (jex == nullptr) {
-                    promise->setValue(RESULT::Boxed::toCpp(jniEnv, jres));
-                } else {
-                    const auto& throwableJniInfo = JniClass<ThrowableJniInfo>::get();
-                    LocalRef<jstring> jmsg(jniEnv, static_cast<jstring>(jniEnv->CallObjectMethod(jex,
-                                                                                                 throwableJniInfo.method_get_message,
-                                                                                                 jex)));
-                    std::string msg = jniUTF8FromString(jniEnv, jmsg.get());
-                    try {
-                        throw std::runtime_error(msg);
-                    } catch (std::exception&) {
-                        promise->setException(std::current_exception());
-                    }
-                }
-            };
-
-            const auto& nativeFutureHandlerJniInfo = JniClass<NativeFutureHandlerJniInfo>::get();
-            auto handler = LocalRef<jobject>(jniEnv, jniEnv->NewObject(nativeFutureHandlerJniInfo.clazz.get(),
-                                                                       nativeFutureHandlerJniInfo.constructor,
-                                                                       reinterpret_cast<jlong>(FutureHandler),
-                                                                       reinterpret_cast<jlong>(p.release())));
-            jniExceptionCheck(jniEnv);
-            const auto& futureJniInfo = JniClass<FutureJniInfo>::get();
-            jniEnv->CallObjectMethod(j, futureJniInfo.method_then, handler.get());
-            jniExceptionCheck(jniEnv);
-            return f;
-        }
-
-        static LocalRef<JniType> fromCpp(JNIEnv* jniEnv, CppType c)
-        {
-            const auto& promiseJniInfo = JniClass<PromiseJniInfo>::get();
-
-            auto promise = std::make_shared<PromiseHolder>();
-            promise->set(jniEnv, jniEnv->NewObject(promiseJniInfo.clazz.get(), promiseJniInfo.constructor));
-            auto future = LocalRef<jobject>(jniEnv, jniEnv->CallObjectMethod(promise->get(), promiseJniInfo.method_get_future));
-            jniExceptionCheck(jniEnv);
-                        
-            c.then([promise, &promiseJniInfo] (Future<CppResType> cppFuture) {
-                JNIEnv* jniEnv = jniGetThreadEnv();
-                try {
-                    auto res = cppFuture.get();
-                    jniEnv->CallVoidMethod(promise->get(), promiseJniInfo.method_set_value, RESULT::Boxed::fromCpp(jniEnv, res).get());
-                } catch (std::exception& e) {
-                    // create a java exception object
-                    const auto& exceptionJniInfo = JniClass<RuntimeExceptionJniInfo>::get();
-                    LocalRef<jobject> jex(jniEnv, jniEnv->NewObject(exceptionJniInfo.clazz.get(), exceptionJniInfo.constructor, String::fromCpp(jniEnv, e.what()).get()));
-                    // call setException()
-                    jniEnv->CallVoidMethod(promise->get(), promiseJniInfo.method_set_exception, jex.get());
-                }
-                jniExceptionCheck(jniEnv);
-            });
-            
-            return future;
         }
     };
 } // namespace djinni
