+++ test-suite/djinni/test.djinni
@@ -1,3 +1,5 @@
+@extern "../../support-lib/future.yaml"
+
 # Helper methods used by various different tests.
 # (Second line of multi-line documentation.
 #   Indented third line of multi-line documentation.)
