+++ test-suite/generated-src/java/com/dropbox/djinni/test/TestHelpers.java
@@ -8,7 +8,6 @@ import java.util.HashMap;
 import java.util.concurrent.atomic.AtomicBoolean;
 import javax.annotation.CheckForNull;
 import javax.annotation.Nonnull;
-import com.snapchat.djinni.Future;
 
 /**
  * Helper methods used by various different tests.
@@ -85,9 +84,10 @@ public abstract class TestHelpers {
     public static native byte[] idBinary(@Nonnull byte[] b);
 
     @Nonnull
-    public static native Future<Integer> getAsyncResult();
+    public static native com.snapchat.djinni.Future<Integer> getAsyncResult();
 
-    public static native Future<String> futureRoundtrip(Future<Integer> f);
+    @Nonnull
+    public static native com.snapchat.djinni.Future<String> futureRoundtrip(@Nonnull com.snapchat.djinni.Future<Integer> f);
 
     public static final class CppProxy extends TestHelpers
     {
