+++ test-suite/generated-src/jni/NativeTestHelpers.cpp
@@ -2,6 +2,7 @@
 // This file generated by Djinni from test.djinni
 
 #include "NativeTestHelpers.hpp"  // my header
+#include "Future_jni.hpp"
 #include "Marshal.hpp"
 #include "NativeAssortedPrimitives.hpp"
 #include "NativeClientInterface.hpp"
@@ -220,19 +221,19 @@ CJNIEXPORT jbyteArray JNICALL Java_com_dropbox_djinni_test_TestHelpers_idBinary(
     } JNI_TRANSLATE_EXCEPTIONS_RETURN(jniEnv, 0 /* value doesn't matter */)
 }
 
-CJNIEXPORT jobject JNICALL Java_com_dropbox_djinni_test_TestHelpers_getAsyncResult(JNIEnv* jniEnv, jobject /*this*/)
+CJNIEXPORT ::snapchat::djinni::FutureAdaptor<::djinni::I32>::JniType JNICALL Java_com_dropbox_djinni_test_TestHelpers_getAsyncResult(JNIEnv* jniEnv, jobject /*this*/)
 {
     try {
         auto r = ::testsuite::TestHelpers::get_async_result();
-        return ::djinni::release(::djinni::FutureAdaptor<::djinni::I32>::fromCpp(jniEnv, std::move(r)));
+        return ::djinni::release(::snapchat::djinni::FutureAdaptor<::djinni::I32>::fromCpp(jniEnv, r));
     } JNI_TRANSLATE_EXCEPTIONS_RETURN(jniEnv, 0 /* value doesn't matter */)
 }
 
-CJNIEXPORT jobject JNICALL Java_com_dropbox_djinni_test_TestHelpers_futureRoundtrip(JNIEnv* jniEnv, jobject /*this*/, jobject f)
+CJNIEXPORT ::snapchat::djinni::FutureAdaptor<::djinni::String>::JniType JNICALL Java_com_dropbox_djinni_test_TestHelpers_futureRoundtrip(JNIEnv* jniEnv, jobject /*this*/, ::snapchat::djinni::FutureAdaptor<::djinni::I32>::JniType j_f)
 {
     try {
-        auto r = ::testsuite::TestHelpers::future_roundtrip(::djinni::FutureAdaptor<::djinni::I32>::toCpp(jniEnv, f));
-        return ::djinni::release(::djinni::FutureAdaptor<::djinni::String>::fromCpp(jniEnv, std::move(r)));
+        auto r = ::testsuite::TestHelpers::future_roundtrip(::snapchat::djinni::FutureAdaptor<::djinni::I32>::toCpp(jniEnv, j_f));
+        return ::djinni::release(::snapchat::djinni::FutureAdaptor<::djinni::String>::fromCpp(jniEnv, r));
     } JNI_TRANSLATE_EXCEPTIONS_RETURN(jniEnv, 0 /* value doesn't matter */)
 }
