+++ test-suite/generated-src/wasm/NativeTestHelpers.cpp
@@ -2,6 +2,7 @@
 // This file generated by Djinni from test.djinni
 
 #include "NativeTestHelpers.hpp"  // my header
+#include "Future_wasm.hpp"
 #include "NativeAssortedPrimitives.hpp"
 #include "NativeClientInterface.hpp"
 #include "NativeColor.hpp"
@@ -15,31 +16,6 @@ namespace djinni_generated {
 
 em::val NativeTestHelpers::cppProxyMethods() {
     static const em::val methods = em::val::array(std::vector<std::string> {
-        "getSetRecord",
-        "checkSetRecord",
-        "getPrimitiveList",
-        "checkPrimitiveList",
-        "getNestedCollection",
-        "checkNestedCollection",
-        "getMap",
-        "checkMap",
-        "getEmptyMap",
-        "checkEmptyMap",
-        "getMapListRecord",
-        "checkMapListRecord",
-        "checkClientInterfaceAscii",
-        "checkClientInterfaceNonascii",
-        "checkClientInterfaceArgs",
-        "checkEnumMap",
-        "checkEnum",
-        "tokenId",
-        "createCppToken",
-        "checkCppToken",
-        "cppTokenId",
-        "checkTokenType",
-        "returnNone",
-        "assortedPrimitivesId",
-        "idBinary",
     });
     return methods;
 }
@@ -121,10 +97,10 @@ em::val NativeTestHelpers::id_binary(const em::val& w_b) {
     return ::djinni::Binary::fromCpp(::testsuite::TestHelpers::id_binary(::djinni::Binary::toCpp(w_b)));
 }
 em::val NativeTestHelpers::get_async_result() {
-    return ::djinni::FutureAdaptor<::djinni::I32>::fromCpp(::testsuite::TestHelpers::get_async_result());
+    return ::snapchat::djinni::FutureAdaptor<::djinni::I32>::fromCpp(::testsuite::TestHelpers::get_async_result());
 }
-em::val NativeTestHelpers::future_roundtrip(const em::val& w_i) {
-    return ::djinni::FutureAdaptor<::djinni::String>::fromCpp(::testsuite::TestHelpers::future_roundtrip(::djinni::FutureAdaptor<::djinni::I32>::toCpp(w_i)));
+em::val NativeTestHelpers::future_roundtrip(const em::val& w_f) {
+    return ::snapchat::djinni::FutureAdaptor<::djinni::String>::fromCpp(::testsuite::TestHelpers::future_roundtrip(::snapchat::djinni::FutureAdaptor<::djinni::I32>::toCpp(w_f)));
 }
 
 EMSCRIPTEN_BINDINGS(test_helpers) {
