+++ test-suite/generated-src/wasm/NativeTestHelpers.hpp
@@ -46,7 +46,8 @@ struct NativeTestHelpers : ::djinni::JsInterface<::testsuite::TestHelpers, Nativ
     static em::val assorted_primitives_id(const em::val& w_i);
     static em::val id_binary(const em::val& w_b);
     static em::val get_async_result();
-    static em::val future_roundtrip(const em::val& w_i);
+    static em::val future_roundtrip(const em::val& w_f);
+
 };
 
 }  // namespace djinni_generated
